# src/TsutaAI/backend-api/openapi.yaml
openapi: 3.0.3
info:
  title: TsutaAI Backend API
  description: |
    TsutaAI プロジェクト・タスク管理システム バックエンドAPI

    ## 認証について
    ほとんどのエンドポイントはJWT認証が必要です。
    `/auth/login`でトークンを取得し、Authorizationヘッダーに`Bearer {token}`形式で含めてください。

    ## AI機能について
    本システムはGroq APIを使用したAI機能を提供します。
    詳細は[AI利用開示](/docs/AI-USAGE-DISCLOSURE.md)をご確認ください。
  version: 1.0.0
  contact:
    name: TsutaAI Support
  license:
    name: Proprietary
    url: https://github.com/yourusername/TsutaAI/blob/main/LICENSE
servers:
  - url: http://localhost:3001
    description: ローカル開発環境
  - url: https://localhost:3443
    description: ローカル開発環境 (HTTPS)
  - url: https://api.tsutaai.example.com
    description: 本番環境
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT トークンによる認証
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        projectId:
          type: integer
          example: 1
        title:
          type: string
          example: デザインモックアップ作成
        description:
          type: string
          example: トップページのデザインモックアップを作成
        status:
          type: string
          enum: [todo, in_progress, review, done, cancelled]
          example: in_progress
        priority:
          type: string
          enum: [low, medium, high, urgent]
          example: high
        assigneeId:
          type: integer
          nullable: true
          example: 2
        estimatedHours:
          type: number
          format: float
          example: 8.5
        actualHours:
          type: number
          format: float
          example: 6.0
        dueDate:
          type: string
          format: date
          nullable: true
          example: "2025-01-15"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /tasks:
    get:
      tags:
        - tasks
      summary: タスク一覧取得
      description: タスク一覧を取得（フィルタ・ページネーション対応）
      operationId: getTasks
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
          description: プロジェクトIDでフィルタ
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, in_progress, review, done, cancelled]
        - name: assigneeId
          in: query
          schema:
            type: integer
          description: 担当者IDでフィルタ
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: タスク一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer

    post:
      tags:
        - tasks
      summary: タスク作成
      description: 新規タスクを作成
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - projectId
                - title
              properties:
                projectId:
                  type: integer
                  example: 1
                title:
                  type: string
                  example: 新しいタスク
                description:
                  type: string
                status:
                  type: string
                  enum: [todo, in_progress, review, done, cancelled]
                  default: todo
                priority:
                  type: string
                  enum: [low, medium, high, urgent]
                  default: medium
                assigneeId:
                  type: integer
                  nullable: true
                estimatedHours:
                  type: number
                  format: float
                dueDate:
                  type: string
                  format: date
      responses:
        '201':
          description: タスク作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  task:
                    $ref: '#/components/schemas/Task'

  # ========================================
  # AI機能エンドポイント
  # ========================================
...